Use sample_app not sample_app2 for all testing
