+++ sample_app/src/test/java/com/mycompany/app/CalculatorMultiplicationTest.java
@@ -1,0 +1,19 @@
+package com.mycompany.app;
+
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertTrue;
+
+import org.junit.Test;
+
+/**
+ * Unit test for calculator multiplication
+ */
+public class CalculatorMultiplicationTest 
+{
+    @Test
+    public void shouldEqualTwentyOne()
+    {
+        float correctAnswer = 21;
+        assertEquals(correctAnswer, Calculator.Multiplication(7, 3), 0.000001);
+    }
+}
